requirements.txt
# MIMIC-CXR Radiology Report Project Requirements
# Python 3.10 environment setup

# Core ML/AI packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers==4.41.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
peft>=0.4.0

# Data processing
pandas>=1.5.0
scikit-learn>=1.2.0
numpy>=1.24.0
duckdb>=1.0.0

# Text processing
sentencepiece>=0.1.99
tiktoken>=0.4.0

# Web interface
streamlit>=1.25.0

# Additional utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=9.5.0
pyyaml>=6.0.0

# Development tools
jupyter>=1.0.0
ipython>=8.0.0

# LLaVA-Med custom architecture support (for microsoft/llava-med-v1.5-mistral-7b)
# Install after other requirements:
# pip install git+https://github.com/microsoft/LLaVA-Med.git
